correction du role , implementation des methodes et de la photo avec l'api flickr nb: il reste 3 controller a implementet competition evenement rencontre!!

+++ src/main/java/greta/federation/dto/LigneCommandeDto.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 public class LigneCommandeDto {
     private Integer id;
 
-    private int quantite;
+    private BigDecimal quantite;
 
 
     private BigDecimal prixUnitaire;
