+++ src/main/java/greta/federation/repository/PosteRepository.java
@@ -4,5 +4,5 @@ import greta.federation.entity.Actualite;
 import greta.federation.entity.Poste;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PosteRepository  extends JpaRepository<Poste, Long> {
+public interface PosteRepository  extends JpaRepository<Poste, Integer> {
 }
